Moved weights_conversion utilities to its own directory to clean the actual user applications in the module
